方案分享/赛题四/队伍名字不能为空/docker_image/app/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

akshare==1.17.22
beautifulsoup4==4.13.4
dashscope==1.23.7
langchain==0.3.26
langchain_community==0.3.27
langchain-community==0.3.27
langchain-core==0.3.68
langchain-huggingface==0.3.0
langchain-mcp-adapters==0.1.9
langchain-milvus==0.2.1
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langgraph==0.5.1
langgraph-supervisor==0.0.27
langchain_milvus==0.2.1
matplotlib==3.9.1
mcp==1.11.0
mplfinance==0.12.10b0
numpy==1.26.4
pandas==2.2.3
pdfplumber==0.11.7
pillow==11.2.1
progressbar2==4.5.0
pyautogen==0.9.0
pydantic==2.11.5
PyExecJS==1.5.1
pymilvus==2.5.12
pysnowball==0.1.5
python-dateutil==2.8.2
python-docx==1.2.0
requests==2.32.4
seaborn==0.13.2
selenium==4.34.2
typing_extensions==4.14.0
webdriver-manager==4.0.2
